move useIsEthFlow to swap/hooks
